Introduce IP KotlinDSL tests for composite builds
